moving userPinMap.py back
